// src/utils/types/SaleDataRowResponseDto.ts
type SaleDataRowResponseDto = {
  representativeId: string;
  enterpriseName: string;
  jan: number | null;
  feb: number | null;
  mar: number | null;
  apr: number | null;
  mai: number | null;
  jun: number | null;
  jul: number | null;
  aug: number | null;
  sep: number | null;
  oct: number | null;
  nov: number | null;
  dec: number | null;
  total: number | null;
};

export default SaleDataRowResponseDto;